backend/requirements.txt
Django==3.2.21
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
dj-database-url==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
whitenoise==6.6.0